renderer: enable anisotropic filtering in renderer1, refs #1530

+++ src/renderercommon/tr_common.h
@@ -107,7 +107,7 @@ void R_NoiseInit(void);
 // If you release a stand-alone game and your mod uses tr_types.h from this build you can safely move them to
 // the glconfig_t struct.
 extern qboolean textureFilterAnisotropic;
-extern int      maxAnisotropy;
+extern float    maxAnisotropy;
 
 // cvars used by both renderers
